Made the progress bar head in round shape Moved fetch course credit code to the course list fragment
